+++ src/components/dialogs/BlockingAnnouncements/index.tsx
@@ -2,7 +2,7 @@ import {View} from 'react-native'
 
 import {Nux} from '#/state/queries/nuxs'
 import {atoms as a} from '#/alf'
-import * as PolicyUpdate20250801 from '#/components/dialogs/BlockingAnnouncements/PolicyUpdate20250801'
+import {Announcement} from '#/components/dialogs/BlockingAnnouncements/announcements/PolicyUpdate202508'
 import {useAnnouncementState} from '#/components/dialogs/BlockingAnnouncements/useAnnouncementState'
 import {createPortalGroup} from '#/components/Portal'
 
@@ -14,7 +14,7 @@ export const Outlet = portalGroup.Outlet
 
 export function BlockingAnnouncements() {
   const state = useAnnouncementState({
-    id: Nux.BlockingAnnouncementPolicyUpdate20250801,
+    id: Nux.BlockingAnnouncementPolicyUpdate202508,
   })
 
   /*
@@ -27,7 +27,7 @@ export function BlockingAnnouncements() {
   return (
     <Portal>
       <View style={[a.fixed, a.inset_0, {zIndex: 9999}]}>
-        <PolicyUpdate20250801.Announcement state={state} />
+        <Announcement state={state} />
       </View>
     </Portal>
   )
